requirements.txt
# os
streamlit
openai
python-dotenv
langchain
chromadb
tiktoken